+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # UNRELEASED
 
+* Always initiate Web Key Directory for new domains
 * Improve style of pasted OpenPGP key (Thanks trashcan55)
 
 # 2.7.4 (2020.10.27)